chore: Various Sonar fixes

+++ cepaslib/src/main/java/com/itachi1706/cepaslib/card/cepas/CEPASProtocol.java
@@ -125,7 +125,7 @@ class CEPASProtocol {
             }
 
             throw new CEPASException("Got generic invalid response: "
-                    + Integer.toHexString(((int) recvBuffer[recvBuffer.length - 2]) & 0xff));
+                    + Integer.toHexString((recvBuffer[recvBuffer.length - 2]) & 0xff));
         }
 
         output.write(recvBuffer, 0, recvBuffer.length - 2);
